Added modal component and reorganized file structure
