+++ src/settings/components/common/Modal.module.scss
@@ -1,0 +1,14 @@
+.modal-wrapper {
+  background-color: rgba(0, 0, 0, 0.5);
+  height: 100vh;
+  left: 0;
+  position: absolute;
+  top: 0;
+  width: 100vw;
+}
+
+.modal {
+  background-color: white;
+  margin: 10vh auto;
+  width: 50vw;
+}
