+++ src/settings/components/entries/EditModal.module.scss
@@ -1,0 +1,6 @@
+.edit-form {
+  display: block;
+  margin: 0 auto;
+  padding: 100px 0;
+  width: 75%;
+}
