+++ src/settings/components/entries/ProfileManager.module.scss
