+++ src/settings/global.d.ts
